% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_dataset_files}
\alias{add_dataset_files}
\title{add_dataset_files}
\usage{
add_dataset_files(
  dataset_DOI,
  paths,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataset_DOI}{A character string representing the DOI of the dataset to which files will be added.}

\item{paths}{A named list where the names correspond to the directory labels and the values are paths to the files to be uploaded.}

\item{BASE_URL}{A character string specifying the base URL of the Dataverse instance. Defaults to the value of the \code{BASE_URL} environment variable.}

\item{API_TOKEN}{A character string representing the API token for authenticating with the Dataverse API. Defaults to the value of the \code{API_TOKEN} environment variable.}
}
\value{
A character vector containing the paths of the files that could not be uploaded, if any.
}
\description{
This function uploads files to a specified Dataverse dataset using its DOI. The function sends HTTP POST requests to the Dataverse API to add files, and reports the status of each upload. If a file cannot be uploaded, it is added to the \code{not_added} list, which is returned at the end.
}
\examples{
# Add files to a dataset with a specific DOI
not_uploaded = add_dataset_files("doi:10.1234/abcde", c("file1.txt", "file2.csv"))

# Add files to a dataset with directory labels
paths = list("directory1" = "file1.txt", "directory2" = "file2.csv")
not_uploaded = add_dataset_files("doi:10.1234/abcde", paths)
}
