% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{convert_datasets_search_to_tibble}
\alias{convert_datasets_search_to_tibble}
\title{convert_datasets_search_to_tibble}
\usage{
convert_datasets_search_to_tibble(datasets_search)
}
\description{
convert_datasets_search_to_tibble
}
