% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{convert_DOI_to_ID}
\alias{convert_DOI_to_ID}
\title{convert_DOI_to_ID}
\usage{
convert_DOI_to_ID(
  dataset_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataset_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of dataset IDs corresponding to the DOIs.
}
\description{
Convert a dataset DOI (persistentId) to the internal Dataverse numeric dataset ID.
}
\examples{
\dontrun{
dotenv::load_dot_env()
convert_DOI_to_ID("doi:10.57745/XXXXXX")
}
}
