% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{convert_metadata}
\alias{convert_metadata}
\title{convert_metadata}
\usage{
convert_metadata(metadata, metadata_yml_path)
}
\arguments{
\item{metadata}{metadata}

\item{metadata_yml_path}{A character string for the name of the metadata R file to write. By default, "metadata_tmp".}
}
\description{
convert_metadata
}
\seealso{
\itemize{
\item \href{https://github.com/super-lou/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/super-lou/dataverseuR}
}
}
