% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{convert_metadata_to_yml}
\alias{convert_metadata_to_yml}
\title{convert_metadata_to_yml}
\usage{
convert_metadata_to_yml(metadata_json_path, overwrite = FALSE)
}
\arguments{
\item{metadata_json_path}{A character vector specifying the paths from which the JSON metadata files should be converted.}

\item{overwrite}{A logical value indicating whether to overwrite an existing metadata file. Defaults to FALSE.}
}
\value{
A vector of character strings containing the YAML metadata paths.
}
\description{
Convert a metadata JSON file from the dataverse to a more user-friendly YAML metadata file. See \code{\link[=get_datasets_metadata]{get_datasets_metadata()}} to retrieve metadata JSON file from a dataverse deposit.
}
\examples{
\dontrun{
metadata_yml_path = convert_metadata_to_yml("metadata.json")
}
}
\seealso{
\itemize{
\item \href{https://github.com/lou-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/lou-heraut/dataverseuR}
}
}
