% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_dataset}
\alias{create_dataset}
\title{create_dataset}
\usage{
create_dataset(
  dataverse,
  metadata_path,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataverse}{A character string specifying the name of the Dataverse in which to create the dataset.}

\item{metadata_path}{A character string representing the path to a JSON file containing the dataset metadata.}

\item{BASE_URL}{A character string representing the base URL for the Dataverse installation (default is the environment variable \code{BASE_URL}).}

\item{API_TOKEN}{A character string representing the API token for authenticating with the Dataverse API (default is the environment variable \code{API_TOKEN}).}
}
\value{
The function returns the DOI of the newly created dataset as a character string.
}
\description{
This function creates a new dataset in a specified Dataverse. It takes metadata from a JSON file, sends a request to the Dataverse API to create the dataset, and returns the persistent DOI of the newly created dataset. If the request fails, the function provides an error message with status code and response content.
}
\examples{
# Create a new dataset using a metadata file
create_dataset("my_dataverse", "path/to/metadata.json")

# Create a new dataset with custom API token and base URL
create_dataset("my_dataverse", "path/to/metadata.json", BASE_URL="https://mydataverse.example.com", API_TOKEN="your_api_token")
}
