% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverseuR.R
\name{dataverseuR}
\alias{dataverseuR}
\title{dataverseuR}
\description{
A dataverse API wraper to enhance deposit procedure with user-friendly YAML parameterization files.
}
\details{
See \href{https://github.com/lou-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/lou-heraut/dataverseuR} for an easier approach.
}
\seealso{
To start :
\itemize{
\item \code{\link[=create_dotenv]{create_dotenv()}} for creating the environment file for credential
}

For metadata generation :
\itemize{
\item \code{\link[=initialise_metadata]{initialise_metadata()}} to get a simple YAML metadata template file
\item \code{\link[=convert_metadata_to_yml]{convert_metadata_to_yml()}} to convert a JSON metadata file from the dataverse to a simplier YAML metadata file
\item \code{\link[=generate_metadata_json]{generate_metadata_json()}} to generate a JSON metadata file for the dataverse from a simplier YAML metadata file
}

For general API actions :
\itemize{
\item \code{\link[=search_datasets]{search_datasets()}} to perform a search on dataverse
\item \code{\link[=create_datasets]{create_datasets()}} to create datasets with a metadata JSON file
\item \code{\link[=modify_datasets]{modify_datasets()}} to modify datasets metadata with a metadata JSON file
\item \code{\link[=add_datasets_files]{add_datasets_files()}} to add files to datasets
\item \code{\link[=delete_datasets_files]{delete_datasets_files()}} to delete files from datasets
\item \code{\link[=delete_all_datasets_files]{delete_all_datasets_files()}} to delete all files from datasets
\item \code{\link[=publish_datasets]{publish_datasets()}} to publish datasets
\item \code{\link[=delete_datasets]{delete_datasets()}} to delete datasets
}

For information about datasets :
\itemize{
\item \code{\link[=get_datasets_metadata]{get_datasets_metadata()}} to get the metadata list of datasets
\item \code{\link[=list_datasets_files]{list_datasets_files()}} to list files of datasets
\item \code{\link[=download_datasets_files]{download_datasets_files()}} to download files of datasets
\item \code{\link[=get_datasets_size]{get_datasets_size()}} to get the size of datasets
\item \code{\link[=get_datasets_metrics]{get_datasets_metrics()}} to get the metrics about datasets
}

Others :
\itemize{
\item \code{\link[=convert_DOI_to_URL]{convert_DOI_to_URL()}} to convert a DOI to an URL
}
}
\author{
Main Developer and Maintainer
\itemize{
\item Louis Héraut \href{mailto:louis.heraut@inrae.fr}{louis.heraut@inrae.fr} INRAE, UR RiverLy, Villeurbanne, France
ORCID : 0009-0006-4372-0923
}
}
