% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_dataset}
\alias{delete_dataset}
\title{delete_dataset}
\usage{
delete_dataset(
  dataset_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataset_DOI}{A character string representing the DOI (Digital Object Identifier) of the dataset to be deleted.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse installation. Default is fetched from the \code{BASE_URL} environment variable.}

\item{API_TOKEN}{A character string representing the API token used for authentication. Default is fetched from the \code{API_TOKEN} environment variable.}
}
\value{
The function does not return any value. It prints messages indicating whether the dataset was successfully deleted or if an error occurred.
}
\description{
This function deletes a dataset in Dataverse using its DOI. It sends a DELETE request to the Dataverse API, passing the DOI of the dataset. Upon success, it prints a success message; otherwise, it reports an error.
}
\examples{
# Delete a dataset with a specific DOI
delete_dataset("doi:10.1234/abcd")

# Delete a dataset with custom parameters
delete_dataset("doi:10.1234/abcd", BASE_URL = "https://dataverse.example.com", API_TOKEN = "your_api_token")
}
