% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_dataset_files}
\alias{delete_dataset_files}
\title{delete_dataset_files}
\usage{
delete_dataset_files(
  dataset_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataset_DOI}{A character string representing the DOI (Digital Object Identifier) of the dataset from which the files will be deleted.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse installation. Default is fetched from the \code{BASE_URL} environment variable.}

\item{API_TOKEN}{A character string representing the API token used for authentication. Default is fetched from the \code{API_TOKEN} environment variable.}
}
\value{
The function does not return any value. It prints messages indicating the status of each file deletion.
}
\description{
This function deletes all files associated with a dataset in the Dataverse repository. It first retrieves the list of files from the specified dataset and then iterates through each file, sending a DELETE request to remove each one. The function reports success or failure for each file.
}
\examples{
# Delete files from a dataset with a specific DOI
delete_dataset_files("doi:10.1234/abcd")

# Delete files from a dataset using custom URL and API token
delete_dataset_files("doi:10.1234/abcd", BASE_URL = "https://dataverse.example.com", API_TOKEN = "your_api_token")
}
