% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_file}
\alias{delete_file}
\title{delete_file}
\usage{
delete_file(
  file_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{file_DOI}{A character string representing the DOI (Digital Object Identifier) of the file that will be deleted.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse installation. Default is fetched from the \code{BASE_URL} environment variable.}

\item{API_TOKEN}{A character string representing the API token used for authentication. Default is fetched from the \code{API_TOKEN} environment variable.}
}
\value{
The function does not return any value. It prints messages indicating the status of the file deletion.
}
\description{
.
}
\examples{
.
}
