% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_file}
\alias{download_file}
\title{download_file}
\usage{
download_file(
  file_DOI,
  save_path,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{file_DOI}{A character string representing the DOI of the file to be downloaded.}

\item{save_path}{A character string representing the local path where the file should be saved.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse repository.
Default is the value of the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string representing the API token for authentication.
Default is the value of the environment variable \code{API_TOKEN}.}
}
\value{
The function downloads the file and saves it to the specified location. If the request fails, an error is returned.
}
\description{
This function downloads a file from a Dataverse repository using its DOI (Digital Object Identifier).
It makes an API call to the Dataverse server and saves the file locally.
}
\examples{
# Download a file using its DOI
download_file("doi:10.5072/FK2/J8SJZB", "downloaded_file.txt")

# Using custom BASE_URL and API_TOKEN
download_file("doi:10.5072/FK2/J8SJZB", "data.csv",
              BASE_URL="https://dataverse.example.com",
              API_TOKEN="your_api_token")
}
