% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{generate_metadata}
\alias{generate_metadata}
\title{generate_metadata}
\usage{
generate_metadata(metadata_yml_path, overwrite_metadata = TRUE)
}
\arguments{
\item{overwrite_metadata}{A logical value indicating whether to overwrite an existing metadata file. Defaults to TRUE.}
}
\value{
A list containing two elements :
\itemize{
\item \code{metadata_path} the path to the generated metadata file and
\item \code{json} the json list equivalent content of the metadata
}
}
\description{
Write a metadata json file and return the associated metadata list based on the R variables contained in the previously initialised R variable metadata environment. See \code{\link[=initialise_metadata]{initialise_metadata()}} to create a new R variable environment.
}
\seealso{
\itemize{
\item \href{https://github.com/super-lou/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/super-lou/dataverseuR}
}
}
