% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{generate_metadata_json}
\alias{generate_metadata_json}
\title{generate_metadata_json}
\usage{
generate_metadata_json(metadata_yml_path, overwrite = FALSE)
}
\arguments{
\item{metadata_yml_path}{A character vector specifying the paths from which the YAML metadata should be read.}

\item{overwrite}{A logical value indicating whether to overwrite an existing metadata file. Defaults to FALSE.}
}
\value{
A vector of character strings containing the JSON metadata paths.
}
\description{
Generate from a YAML metadata file a JSON metadata file used by dataverse to create or modify a dataset. See \code{\link[=initialise_metadata]{initialise_metadata()}} or \code{\link[=get_datasets_metadata]{get_datasets_metadata()}} and \code{\link[=convert_metadata_to_yml]{convert_metadata_to_yml()}} to get a metadata YAML file.
}
\examples{
\dontrun{
# First option 
# Get a YAML file from a template
initialise_metadata("metadata.yml")
# Customize the YAML file, then generate the JSON equivalent
metadata_json_path = generate_metadata_json("metadata.yml")

# Second option
# Get a YAML file from a dataverse deposit
dataset_DOI = "doi:10.57745/LNBEGZ"
get_datasets_metadata(dataset_DOI, "metadata.json")
convert_metadata_to_yml("metadata.json")
# Customize the YAML file, then generate the JSON equivalent
metadata_json_path = generate_metadata_json("metadata.yml")
}
}
\seealso{
\itemize{
\item \href{https://github.com/lou-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/lou-heraut/dataverseuR}
}
}
