% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dataset_metadata}
\alias{get_dataset_metadata}
\title{get_dataset_metadata}
\usage{
get_dataset_metadata(
  dataset_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataset_DOI}{A character string representing the DOI of the dataset whose metadata is to be retrieved.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse repository. The default is the value of the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string representing the API token for authentication. The default is the value of the environment variable \code{API_TOKEN}.}
}
\value{
A list containing the dataset metadata if the request is successful. If the request fails, the function stops and returns an error message.
}
\description{
This function retrieves metadata for a dataset using its DOI (Digital Object Identifier) from a Dataverse repository. It makes an API call to the Dataverse server using the provided or default credentials (BASE_URL and API_TOKEN), and returns the dataset's metadata if the request is successful. If the request fails, it prints an error message with the status code and response content.
}
\examples{
# Retrieve metadata for a dataset using its DOI
dataset_info = get_dataset_metadata("doi:10.1234/abcde12345")

# Using custom BASE_URL and API_TOKEN
dataset_info = get_dataset_metadata("doi:10.1234/abcde12345", BASE_URL="https://dataverse.example.com", API_TOKEN="your_api_token")
}
