% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets_size}
\alias{get_datasets_size}
\title{get_datasets_size}
\usage{
get_datasets_size(
  dataset_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataset_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\value{
A tibble containing datasets DOI and their corresponding total storage sizes in bytes.
}
\description{
Retrieves the total storage size (in bytes) of a selection of datasets.
}
\examples{
\dontrun{
dotenv::load_dot_env()
dataset_DOI = "doi:10.57745/LNBEGZ"
get_datasets_size(dataset_DOI)
}
}
\seealso{
\itemize{
\item \href{https://github.com/lou-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/lou-heraut/dataverseuR}
\item \href{https://github.com/lou-heraut/dataverseuR_toolbox/blob/main/RiverLy_HCERES_2025/script.R}{R example in context} \url{https://github.com/lou-heraut/dataverseuR_toolbox/blob/main/RiverLy_HCERES_2025/script.R}
\item \href{https://guides.dataverse.org/en/5.3/api/native-api.html#report-the-data-file-size-of-a-dataverse}{Native API documentation} \url{https://guides.dataverse.org/en/5.3/api/native-api.html#report-the-data-file-size-of-a-dataverse}
}
}
