% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dotenv}
\alias{get_dotenv}
\title{get_dotenv}
\usage{
get_dotenv(dotenv_path = file.path(getwd(), "dist.env"))
}
\arguments{
\item{dotenv_path}{A character string representing the destination path where the \code{.env} file should be copied. By default, it copies to the \code{dist.env} file in the working directory.}
}
\value{
The function returns the path to the copied \code{.env} file.
}
\description{
This function copies a predefined \code{.env} file from the package's extdata directory to a specified path. It also prints a message advising the user to fill in their credentials and rename the file to \code{.env}. The copied file serves as a template for environment variables required for authentication or other configuration.
}
\examples{
# Copy the .env file to a specific location
get_dotenv("path/to/your/.env")

# Copy the .env file to the default location (current working directory)
get_dotenv()
}
