% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_dataset_files}
\alias{list_dataset_files}
\title{list_dataset_files}
\usage{
list_dataset_files(
  dataset_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataset_DOI}{A character string representing the DOI (Digital Object Identifier) of the dataset for which the files are to be listed.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse API (default is taken from the system environment variable \code{BASE_URL}).}

\item{API_TOKEN}{A character string representing the API token used for authentication with the Dataverse API (default is taken from the system environment variable \code{API_TOKEN}).}
}
\value{
A data frame containing information about the files associated with the dataset, excluding the description field.
}
\description{
This function retrieves the list of files associated with a dataset from a Dataverse repository, given the dataset's DOI. It makes an API request to the Dataverse server and processes the response to extract file information, such as file names and metadata. The function uses the base URL and API token provided by the user or defaults to system environment variables.
}
\examples{
# List files for a dataset with a specific DOI
files <- list_dataset_files("doi:10.1234/abc123")

# Use custom URL and token to list dataset files
files <- list_dataset_files("doi:10.1234/abc123", BASE_URL = "https://dataverse.example.com", API_TOKEN = "your_api_token")
}
