% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{modify_dataset_metadata}
\alias{modify_dataset_metadata}
\title{modify_dataset_metadata}
\usage{
modify_dataset_metadata(
  dataverse,
  dataset_DOI,
  metadata_path,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{dataverse}{A character string representing the name of the Dataverse repository.}

\item{dataset_DOI}{A character string representing the DOI of the dataset to modify.}

\item{metadata_path}{A character string indicating the file path to the metadata JSON file that will be used for the update.}

\item{BASE_URL}{A character string representing the base URL of the Dataverse installation. Defaults to the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string representing the API token for authentication. Defaults to the environment variable \code{API_TOKEN}.}
}
\value{
The function returns the DOI of the modified dataset.
}
\description{
This function modifies the metadata of a dataset in a Dataverse repository. It accepts the dataset DOI, the path to a metadata JSON file, and uses the Dataverse API to update the dataset's metadata. The function checks the response for success and outputs relevant information. It is designed to facilitate metadata management and updating in Dataverse.
}
\examples{
# Modify the metadata of a dataset
modify_dataset_metadata(dataverse = "myDataverse", 
                         dataset_DOI = "doi:10.1234/abcd", 
                         metadata_path = "metadata.json")

# Modify the metadata with a custom BASE_URL and API_TOKEN
modify_dataset_metadata(dataverse = "myDataverse", 
                         dataset_DOI = "doi:10.1234/abcd", 
                         metadata_path = "metadata.json",
                         BASE_URL = "https://mydataverse.org", 
                         API_TOKEN = "myapitoken123")
}
