% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{modify_datasets}
\alias{modify_datasets}
\title{modify_datasets}
\usage{
modify_datasets(
  dataverse,
  dataset_DOI,
  metadata_json_path,
  wait_time = 2,
  n_retries = 3,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataverse}{A character string specifying the name of the dataverse in which to create datasets.}

\item{dataset_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{metadata_json_path}{A vector of character string for the paths of JSON files containing the datasets metadata.}

\item{wait_time}{An integer for the time in seconds to wait between requests. By default, 2 seconds.}

\item{n_retries}{An integer for the maximum number of retries to reach dataverse server. By default, 3 retries.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\description{
Modify datasets metadata from a selection of metadata JSON file. See \code{\link[=generate_metadata_json]{generate_metadata_json()}} in order to generate efficiently metadata JSON file.
}
\examples{
\dontrun{
dotenv::load_dot_env()
dataset_DOI = "doi:10.57745/LNBEGZ"
metadata_json_path = "path/to/read/metadata.json"
modify_datasets("inrae", dataset_DOI, metadata_json_path)
}
}
\seealso{
\itemize{
\item \href{https://github.com/lou-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/lou-heraut/dataverseuR}
\item \href{https://github.com/lou-heraut/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}{R example in context} \url{https://github.com/lou-heraut/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}
\item \href{https://guides.dataverse.org/en/5.3/api/native-api.html#update-metadata-for-a-dataset}{Native API documentation} \url{https://guides.dataverse.org/en/5.3/api/native-api.html#update-metadata-for-a-dataset}
}
}
