% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{move_datasets}
\alias{move_datasets}
\title{move_datasets}
\usage{
move_datasets(
  dataset_DOI,
  target_dataverse,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataset_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{target_dataverse}{A character string giving the alias of the destination Dataverse.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\description{
Move one or more datasets from their current Dataverse collection to another Dataverse collection (superuser or authorized user only).
}
\examples{
\dontrun{
dotenv::load_dot_env()
move_datasets(dataset_ID=12345,
              target_dataverse="new_dataverse")
}
}
