% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_datasets_files}
\alias{add_datasets_files}
\title{add_datasets_files}
\usage{
add_datasets_files(
  dataset_DOI,
  file_paths,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  timeout = 600,
  verbose = TRUE
)
}
\arguments{
\item{dataset_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{file_paths}{A vector of character string of paths of the files that needs to be uploaded. If you want to conserve a directory structure, set the name of each file path to the directory name wanted. If \code{dataset_DOI} contains more than one dataset DOI, \code{file_paths} needs to be a list of vector of character string, each vector of this list will be associated to each dataset identified in \code{dataset_DOI}.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{timeout}{An integer to set the number of second before timeout of the request. By default, 600.}

\item{verbose}{If FALSE, no processing informations are displayed. By default, TRUE.}
}
\value{
A character vector containing the paths of the files that could not be uploaded, if any.
}
\description{
This function uploads files to a dataset.
}
\seealso{
\itemize{
\item \href{https://github.com/super-lou/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/super-lou/dataverseuR}
\item \href{https://github.com/super-lou/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}{R example in context} \url{https://github.com/super-lou/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}
\item \href{https://guides.dataverse.org/en/5.3/api/native-api.html#add-a-file-to-a-dataset}{Native API documentation} \url{https://guides.dataverse.org/en/5.3/api/native-api.html#add-a-file-to-a-dataset}
}
}
