% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{convert_DOI_to_URL}
\alias{convert_DOI_to_URL}
\title{convert_DOI_to_URL}
\usage{
convert_DOI_to_URL(DOI)
}
\arguments{
\item{DOI}{A character string specifying the DOI to convert.}
}
\value{
A character string for the associated URL.
}
\description{
Convert a DOI to a functional URL.
}
\examples{
convert_DOI_to_URL("doi:10.57745/LNBEGZ")
}
\seealso{
\itemize{
\item \href{https://github.com/super-lou/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/super-lou/dataverseuR}
}
}
