% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverseuR.R
\name{convert_metadata}
\alias{convert_metadata}
\title{convert_metadata}
\usage{
convert_metadata(metadata, environment_name = "META")
}
\description{
convert_metadata
}
\seealso{
\itemize{
\item \href{https://github.com/super-lou/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/super-lou/dataverseuR}
}
}
