% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{convert_metadata}
\alias{convert_metadata}
\title{convert_metadata}
\usage{
convert_metadata(metadata, environment_name = "META")
}
\arguments{
\item{metadata}{metadata}

\item{environment_name}{The name of the R variable metadata environment variable containing the metadata values. Defaults to "META".}
}
\description{
convert_metadata
}
\seealso{
\itemize{
\item \href{https://github.com/super-lou/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/super-lou/dataverseuR}
}
}
