% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_datasets_files}
\alias{delete_datasets_files}
\title{delete_datasets_files}
\usage{
delete_datasets_files(
  files_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{files_DOI}{A vector of character string of the DOI of files to be processed.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If FALSE, no processing informations are displayed. By default, TRUE.}
}
\description{
Delete files based on their DOI.
}
\examples{
.
}
