% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_datasets_files}
\alias{download_datasets_files}
\title{download_datasets_files}
\usage{
download_datasets_files(
  dataset_DOI,
  dirpath = "dataverse_files",
  overwrite = FALSE,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataset_DOI}{A vector of character string of dataset DOI(s). All files from each dataset will be downloaded.}

\item{dirpath}{A character string representing the base local directory path where files should be saved. Defaults to \code{"dataverse_files"}. Subdirectories will be created for each dataset based on sanitized dataset DOIs.}

\item{overwrite}{A logical value indicating whether to overwrite existing files. Defaults to FALSE.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\description{
Download all files from one or multiple datasets. Files are automatically organized in subdirectories named after sanitized dataset DOIs.
}
\examples{
\dontrun{
dotenv::load_dot_env()

# Download all files from a single dataset
dataset_DOI = "doi:10.57745/LNBEGZ"
download_datasets_files(dataset_DOI)
# Files saved in: dataverse_files/doi_10-57745_LNBEGZ/

# Download all files from multiple datasets
dataset_DOI = c("doi:10.57745/LNBEGZ", "doi:10.57745/TGYZ8L")
download_datasets_files(dataset_DOI)
# Files saved in: dataverse_files/doi_10-57745_LNBEGZ/ and dataverse_files/doi_10-57745_TGYZ8L/

# With custom save directory
download_datasets_files(dataset_DOI, dirpath="my_files")

# Works with DRAFT datasets too
dataset_DOI = "doi:10.57745/LNBEGZ"
download_datasets_files(dataset_DOI)
}
}
\seealso{
\itemize{
\item \href{https://github.com/louis-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/louis-heraut/dataverseuR}
\item \href{https://github.com/louis-heraut/dataverseuR_toolbox/blob/main/DRYvER/modify_README.R}{R example in context} \url{https://github.com/louis-heraut/dataverseuR_toolbox/blob/main/DRYvER/modify_README.R}
\item \href{https://guides.dataverse.org/en/5.3/api/native-api.html#accessing-downloading-files}{Native API documentation} \url{https://guides.dataverse.org/en/5.3/api/native-api.html#accessing-downloading-files}
}
}
