% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets_metrics}
\alias{get_datasets_metrics}
\title{get_datasets_metrics}
\usage{
get_datasets_metrics(
  datasets_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{datasets_DOI}{A character string or vector representing the DOI of the datasets.}

\item{BASE_URL}{A character string representing the base URL for the Dataverse installation (default is the environment variable \code{BASE_URL}).}

\item{API_TOKEN}{A character string representing the API token for authentication (default is the environment variable \code{API_TOKEN}).}
}
\value{
A named list containing total views, unique views, total downloads, unique downloads, and citations.
}
\description{
This function retrieves dataset metrics (views, downloads, and citations) from the Dataverse API for a given dataset DOI.
}
\examples{
# Retrieve dataset metrics
get_dataset_metrics("doi:10.5072/FK2/J8SJZB")
}
