% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_datasets_size}
\alias{get_datasets_size}
\title{get_datasets_size}
\usage{
get_datasets_size(
  datasets_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{datasets_DOI}{A character string or vector representing the dataset DOIs.}

\item{BASE_URL}{A character string representing the base URL for the Dataverse installation (default is the environment variable \code{BASE_URL}).}

\item{API_TOKEN}{A character string representing the API token for authentication (default is the environment variable \code{API_TOKEN}).}
}
\value{
A tibble containing dataset DOIs and their corresponding total storage sizes in bytes.
}
\description{
This function retrieves the total storage size (in bytes) of a dataset from the Dataverse API.
}
\examples{
# Retrieve dataset storage size
get_datasets_size("24")
}
