% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverseur.R
\name{initialise_metadata}
\alias{initialise_metadata}
\title{initialise_metadata}
\usage{
initialise_metadata(environment_name = "META")
}
\arguments{
\item{environment_name}{A character string representing the name of the environment to be created. The default value is "META".}
}
\value{
The function does not return a value. It creates a new environment and assigns it to the specified variable in the global environment.
}
\description{
This function creates a new environment and assigns it to a specified variable name. By default, the environment is assigned to a variable named "META". This function is useful for initializing metadata storage in a separate environment.
}
\examples{
# Initialize metadata with the default environment name "META"
initialise_metadata()

# Initialize metadata with a custom environment name
initialise_metadata("CustomEnv")
}
