% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{initialise_metadata}
\alias{initialise_metadata}
\title{initialise_metadata}
\usage{
initialise_metadata(metadata_yml_path = "metadata.yml", overwrite = FALSE)
}
\arguments{
\item{metadata_yml_path}{A character string specifying the path to which the YAML metadata template should be written. By default, \code{"metadata.yml"} in the working directory.}

\item{overwrite}{A logical value indicating whether to overwrite an existing metadata template. Defaults to FALSE.}
}
\value{
The function returns the path to the created .env file.
}
\description{
Create an empty example for a dataverse YAML metadata file.
}
\examples{
\dontrun{
# Create the metadata template to the working directory
initialise_metadata()
# or create the metadata template in a remote location
initialise_metadata("path/to/metadata_project.yml")
}
}
\seealso{
\itemize{
\item \href{https://github.com/louis-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/louis-heraut/dataverseuR}
}
}
