% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverseuR.R
\name{initialise_metadata}
\alias{initialise_metadata}
\title{initialise_metadata}
\usage{
initialise_metadata(environment_name = "META")
}
\arguments{
\item{environment_name}{A character string representing the name of the R variable metadata environment to be created. The default value is \code{"META"}.}
}
\description{
Create an empty R variable metadata environment in which all assigned R variables will represent a dataverse metadata for a unique dataset. See \code{\link[=generate_metadata]{generate_metadata()}} to create the associated json metadata file.
}
\examples{
a
}
