% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_datasets_files}
\alias{list_datasets_files}
\title{list_datasets_files}
\usage{
list_datasets_files(
  datasets_DOI,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN")
)
}
\arguments{
\item{datasets_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}
}
\value{
A \link[dplyr:reexports]{tibble} containing files information.
}
\description{
Retrieves files informations of a selection of datasets.
}
\examples{
a
}
