% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{modify_datasets}
\alias{modify_datasets}
\title{modify_datasets}
\usage{
modify_datasets(
  dataverse,
  datasets_DOI,
  metadata_paths,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataverse}{A character string specifying the name of the dataverse in which to create datasets.}

\item{datasets_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{metadata_paths}{A vector of character string for the paths of json files containing the datasets metadata.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If FALSE, no processing informations are displayed. By default, TRUE.}
}
\description{
Modify datasets metadata from a selection of metadata json file. See \code{\link[=generate_metadata]{generate_metadata()}} in order to generate efficiently metadata json file.
}
\examples{
a
}
