% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{publish_datasets}
\alias{publish_datasets}
\title{publish_datasets}
\usage{
publish_datasets(
  datasets_DOI,
  type = "major",
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{datasets_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{type}{A character string specifying the type of publication. Use \code{"major"} if at least a file as been modified of supressed, if only metadata have changed, use \code{"minor"}.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If FALSE, no processing informations are displayed. By default, TRUE.}
}
\description{
Publish a selection of datasets.
}
\examples{
a
}
