% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{publish_datasets}
\alias{publish_datasets}
\title{publish_datasets}
\usage{
publish_datasets(
  dataset_DOI,
  type = "major",
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{dataset_DOI}{A vector of character string representing the DOI of datasets that will be process.}

\item{type}{A character string specifying the type of publication. Use \code{"major"} if at least a file as been modified of supressed, if only metadata have changed, use \code{"minor"}.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\description{
Publish a selection of datasets.
}
\examples{
\dontrun{
dotenv::load_dot_env()
dataset_DOI = "doi:10.57745/LNBEGZ"
publish_datasets(dataset_DOI)
}
}
\seealso{
\itemize{
\item \href{https://github.com/louis-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/louis-heraut/dataverseuR}
\item \href{https://github.com/louis-heraut/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}{R example in context} \url{https://github.com/louis-heraut/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}
\item \href{https://guides.dataverse.org/en/5.3/api/native-api.html#publish-a-dataset}{Native API documentation} \url{https://guides.dataverse.org/en/5.3/api/native-api.html#publish-a-dataset}
}
}
