% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rename_datasets_files}
\alias{rename_datasets_files}
\title{rename_datasets_files}
\usage{
rename_datasets_files(
  file_DOI,
  new_name,
  is_DOI_ID = FALSE,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{file_DOI}{A vector of character string of the DOI of files to be processed.}

\item{new_name}{A vector of character string representing the new filename.}

\item{is_DOI_ID}{If the dataset is not published yet, the DOI of the file does not exist, so \code{file_DOI} needs to be the file id of the database instead of a DOI. So if \code{TRUE}, \code{file_DOI} is \code{id} from the results of \code{\link[=list_datasets_files]{list_datasets_files()}}, elsewhere if \code{FALSE}, \code{file_DOI} is actual file DOI. Default, \code{FALSE}.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If \code{FALSE}, no processing informations are displayed. Defaults to \code{TRUE}.}
}
\description{
Renames a file based on their DOI.
}
\examples{
\dontrun{
dotenv::load_dot_env()

# In general
file_DOI = "doi:10.57745/QB73Q0"
rename_datasets_files(file_DOI, new_name="LICENCE_Etalab.pdf")

# A more complexe example when the dataset is a DRAFT
dataset_DOI = "doi:10.57745/LNBEGZ"
files = list_datasets_files(dataset_DOI)
licence = dplyr::filter(files, grepl("LICENCE", label))
rename_datasets_files(file_DOI=licence$id,
                      new_name="LICENCE_Etalab.pdf",
                      is_DOI_ID=TRUE)
}
}
\seealso{
\itemize{
\item \href{https://github.com/louis-heraut/dataverseuR}{dataverseuR GitHub documentation} \url{https://github.com/louis-heraut/dataverseuR}
\item \href{https://github.com/louis-heraut/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}{R example in context} \url{https://github.com/louis-heraut/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}
\item \href{https://guides.dataverse.org/en/5.3/api/native-api.html#updating-file-metadata}{Native API documentation} \url{https://guides.dataverse.org/en/5.3/api/native-api.html#updating-file-metadata}
}
}
