% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rename_datasets_files}
\alias{rename_datasets_files}
\title{rename_datasets_files}
\usage{
rename_datasets_files(
  file_DOI,
  new_name,
  BASE_URL = Sys.getenv("BASE_URL"),
  API_TOKEN = Sys.getenv("API_TOKEN"),
  verbose = TRUE
)
}
\arguments{
\item{file_DOI}{A vector of character string of the DOI of files to be processed.}

\item{new_name}{A vector of character string representing the new filename.}

\item{BASE_URL}{A character string for the base URL of the Dataverse API. By default, it uses the value from the environment variable \code{BASE_URL}.}

\item{API_TOKEN}{A character string for the API token required to authenticate the request. By default, it uses the value from the environment variable \code{API_TOKEN}.}

\item{verbose}{If FALSE, no processing informations are displayed. By default, TRUE.}
}
\description{
Renames a file based on their DOI.
}
\references{
\href{https://guides.dataverse.org/en/5.3/api/native-api.html#updating-file-metadata}{Native API documentation}
}
\seealso{
\href{https://github.com/super-lou/dataverseuR_toolbox/blob/main/Explore2/post_hydrological_projection.R}{R example in context}
}
